Fix Apache instructions and PHP extension list
For the Apache instruction, the Options directive does exist, but it's
not the one that controls .htaccess usage, that's AllowOverride.

For the PHP module, correct mbstrings to mbstring and remove hash, since
the [hash extension][1] has been included in PHP since 5.1.2, and is core
since PHP 7.4, both of which are above the minimum Friendica version.

[1]: https://www.php.net/manual/en/hash.installation.php